restructure darwin into its own package
