+++ fakenet/diverters/darwin/utils.py
